requirements.txt
# Dependencias principales del proyecto
Django==5.2.7
python-dotenv==1.1.1
whitenoise==6.9.0
gunicorn==23.0.0

# Base de datos y ORM
sqlparse==0.5.3
asgiref==3.9.1
typing_extensions==4.14.1
tzdata==2025.2

# Autenticación y seguridad  
django-allauth==65.12.0
django-axes==8.0.0
django-cors-headers==4.9.0
django-environ==0.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0

# Formularios y widgets
django-formtools==2.5.1
django-widget-tweaks==1.5.0

# Autenticación de dos factores
django-two-factor-auth==1.17.0
django-otp==1.6.1
pyotp==2.9.0
qrcode==7.4.2

# Utilidades
pillow==11.3.0
requests==2.32.5
setuptools==80.9.0

# PostgreSQL y utilidades para DATABASE_URL
dj-database-url
psycopg[binary]
django-storages==1.14.6
boto3
